+++ src/ConnectionInterface.php
@@ -29,7 +29,7 @@ interface ConnectionInterface
 
     public function disconnect(): bool;
 
-    public function useTube(string $tube): self;
+    public function useTube(string $tube): void;
 
     /**
      * @return string|int
@@ -46,29 +46,29 @@ interface ConnectionInterface
     /**
      * @param string|int $id
      */
-    public function touch($id): self;
+    public function touch($id): void;
 
     /**
      * @param string|int $id
      */
-    public function release($id, int $priority = self::DEFAULT_PRIORITY, int $delay = self::DEFAULT_DELAY): self;
+    public function release($id, int $priority = self::DEFAULT_PRIORITY, int $delay = self::DEFAULT_DELAY): void;
 
     /**
      * @param string|int $id
      */
-    public function bury($id, int $priority = self::DEFAULT_PRIORITY): self;
+    public function bury($id, int $priority = self::DEFAULT_PRIORITY): void;
 
     /**
      * @param string|int $id
      */
-    public function delete($id): self;
+    public function delete($id): void;
 
-    public function watch(string $tube): self;
+    public function watch(string $tube): int;
 
     /**
      * @return int|null Number of tubes being watched or null when last tube cannot be ignored
      */
-    public function ignore(string $tube): ?int;
+    public function ignore(string $tube): int;
 
     /**
      * @param string|int $id