+++ tests/PoolTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Phlib\Beanstalk;
 
+use Phlib\Beanstalk\Exception\CommandException;
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
 use Phlib\Beanstalk\Exception\RuntimeException;
 use Phlib\Beanstalk\Pool\Collection;
@@ -77,13 +78,17 @@ class PoolTest extends TestCase
 
     public function testIgnoreDoesNotAllowLessThanOneWatching(): void
     {
+        $this->expectException(CommandException::class);
+        $this->expectExceptionMessage('Cannot ignore the only tube in the watch list');
+
         // 'default' tube is already being watched
-        static::assertNull($this->pool->ignore('default'));
+        $this->pool->ignore('default');
     }
 
     public function testIgnore(): void
     {
-        $this->pool->watch('test-tube');
+        $actual = $this->pool->watch('test-tube');
+        static::assertSame(2, $actual);
         static::assertSame(1, $this->pool->ignore('default'));
     }
 
@@ -582,7 +587,8 @@ class PoolTest extends TestCase
 
     public function testListTubesWatched(): void
     {
-        $this->pool->watch('test');
+        $actual = $this->pool->watch('test');
+        static::assertSame(2, $actual);
         static::assertSame(['default', 'test'], $this->pool->listTubesWatched());
     }